+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/PartyV2/OnlineAsyncTaskAccelByteRefreshV2PartySession.cpp
@@ -66,12 +66,19 @@ void FOnlineAsyncTaskAccelByteRefreshV2PartySession::TriggerDelegates()
 
 	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("bWasSuccessful: %s"), LOG_BOOL_FORMAT(bWasSuccessful));
 
-	if (bWasSuccessful && bWasSessionRemoved)
+	if (bWasSuccessful)
 	{
 		FOnlineSessionV2AccelBytePtr SessionInterface = nullptr;
 		if (ensureAlways(FOnlineSessionV2AccelByte::GetFromSubsystem(Subsystem, SessionInterface)))
 		{
-			SessionInterface->TriggerOnSessionRemovedDelegates(SessionName);
+			if (bWasSessionRemoved)
+			{
+				SessionInterface->TriggerOnSessionRemovedDelegates(SessionName);
+			}
+			else
+			{
+				SessionInterface->TriggerOnSessionUpdateReceivedDelegates(SessionName);
+			}
 		}
 	}